Fixed the analyzePossibleCards method in UnoRuleUtils, moved the utils tests to the respective package, fixed and tweaked the documentation a bit.
