+++ src/test/java/com/github/markozajc/juno/utils/UnoRuleUtilsTest.java
@@ -1,0 +1,16 @@
+package com.github.markozajc.juno.utils;
+
+import static org.junit.jupiter.api.Assertions.fail;
+
+import org.junit.jupiter.api.Test;
+
+
+class UnoRuleUtilsTest {
+
+	@Test
+	void testAnalyzePossibleCards() {
+		fail("TODO");
+		// TODO
+	}
+
+}
